refactor: flatten agent directory structure
